refactor: AIClient를 KaggomAIClient와 ImageAIClient로 분리
