+++ src/main/java/com/greeenai/greeenai/infra/ai/client/KaggomAIClient.java
@@ -2,29 +2,23 @@ package com.greeenai.greeenai.infra.ai.client;
 
 import static com.greeenai.greeenai.global.error.exception.ErrorCode.*;
 
+import com.greeenai.greeenai.global.error.exception.CustomException;
 import com.greeenai.greeenai.infra.ai.dto.request.DiaryEntry;
 import com.greeenai.greeenai.infra.ai.dto.request.GenerateDiaryRequest;
-import com.greeenai.greeenai.infra.ai.dto.request.GenerateImageRequest;
 import com.greeenai.greeenai.infra.ai.dto.request.GenerateQuestionsRequest;
 import com.greeenai.greeenai.infra.ai.dto.response.GenerateDiaryResponse;
 import com.greeenai.greeenai.infra.ai.dto.response.GenerateQuestionsResponse;
 import com.greeenai.greeenai.infra.ai.dto.response.GenerateQuestionsResponse.GeneratedQuestion;
-import com.greeenai.greeenai.infra.ai.dto.response.GeneratedImage;
-import com.greeenai.greeenai.global.error.exception.CustomException;
 import java.util.List;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
 
 @Service
 @RequiredArgsConstructor
-public class AIClient {
+public class KaggomAIClient {
 
     private final WebClient kaggomWebClient;
-    private final WebClient aiWebClient;
 
     public List<GeneratedQuestion> generateQuestions(List<String> imageUrls) {
         GenerateQuestionsRequest request = GenerateQuestionsRequest.from(imageUrls);
@@ -54,28 +48,6 @@ public class AIClient {
         }
 
         throw new CustomException(AI_QUESTION_GENERATION_FAILED);
-    }
-
-    public GeneratedImage generateImage(List<DiaryEntry> entries) {
-        GenerateImageRequest request = GenerateImageRequest.of(entries);
-
-        ResponseEntity<byte[]> response = aiWebClient
-                .post()
-                .uri("/generate-image")
-                .bodyValue(request)
-                .retrieve()
-                .toEntity(byte[].class)
-                .block();
-
-        if (response == null || response.getBody() == null) {
-            throw new CustomException(AI_IMAGE_GENERATION_FAILED);
-        }
-
-        return GeneratedImage.of(
-                response.getBody(),
-                Optional.ofNullable(response.getHeaders().getContentType())
-                        .map(MediaType::toString)
-                        .orElse("image/png"));
     }
 
     public String generateDiary(List<DiaryEntry> entries) {
